Backend/Karthikeya/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.3
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
